Fix error message for parse error on search result page
