+++ gradle/changelog/search_parse_error.yaml
@@ -1,0 +1,2 @@
+- type: Fixed
+  description: Error message for parse error on search result page ([#1768](https://github.com/scm-manager/scm-manager/pull/1768))
